+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## 2.0.1 under development
 
-- Chg #75: Add method `toSnakeCase()` to Inflector class (soodssr)
+- New #75: Add method `Inflector::toSnakeCase()` that convert word to "snake_case" (soodssr)
 
 ## 2.0.0 February 10, 2021